featExt  returns a vector of bin indices instead of raw pow of magnitudes.

+++ src/fft_processor.hpp
@@ -24,7 +24,7 @@ public:
     virtual ~FFTProcessor(void);
 
     std::vector<float> FFT(const std::vector<float>& frame);
-    std::vector<float> featExt(const std::vector<float>& frame);
+    std::vector<int> featExt(const std::vector<float>& frame);
 };
 
 #endif
